pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
pyjwt = "^2.8.0"
python-multipart = "^0.0.9"
django = "^5.1.1"
django-crispy-forms = "^2.3"
llama-index-agent-openai = "^0.4.0"
django-restframework = "^0.0.1"
drf-spectacular = "^0.27.2"
rest-framework-simplejwt = "^0.0.2"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.6.0"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
isort = "^5.13.2"
bandit = "^1.7.7"
mypy = "^1.8.0"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
